make it possible to download files larger than 2GB in tests
Currently, tests a'la `downloadLargeFile` download a file of size 5MB. But now you can out 2048 or even 4096 instead.

+++ src/test/java/integration/server/Delayer.java
@@ -30,7 +30,7 @@ class Delayer {
     }
   }
 
-  static void writeSlowly(OutputStream os, int contentLength, InputStream content, long duration) throws IOException {
+  static void writeSlowly(OutputStream os, long contentLength, InputStream content, long duration) throws IOException {
     Delayer delayer = new Delayer(duration);
     long delay = Math.max(duration / (contentLength + 1), 1);
     delayer.delay(delay);
